requirements.txt
torch
transformers
datasets
pyyaml
matplotlib
jupyter